+++ src/config/options.js
@@ -39,6 +39,7 @@ const defaultOptions = Object.assign({
   nodeModulesDir: 'node_modules',
   pluginsDir: 'plugins',
   bundlesDir: 'src',
+  activitiesDir: 'activities',
   themesDir: 'web/themes',
 
   fontlimit: 1024,
@@ -47,9 +48,8 @@ const defaultOptions = Object.assign({
   imgName: 'img',
   mediaName: 'media',
   copyName: 'img',
-
   isESlint: true,
-
+  cpuNumber: 2,
   isNeedCommonChunk: '.is-need-common-chunk'
   
 }, parameters);
@@ -61,6 +61,7 @@ const nodeModulesDir = path.resolve(rootDir, defaultOptions.nodeModulesDir);
 const pluginsDir = path.resolve(rootDir, defaultOptions.pluginsDir);
 const themesDir = path.resolve(rootDir,defaultOptions.themesDir);
 const bundlesDir = path.resolve(rootDir, defaultOptions.bundlesDir);
+const activitiesDir = path.resolve(rootDir, defaultOptions.activitiesDir);
 
 // 是否编译相应模块
 const isBuildAllModule =  !!specialArgv.module ? false : true;
@@ -114,6 +115,7 @@ const options = Object.assign({}, defaultOptions, {
   pluginsDir,
   themesDir,
   bundlesDir,
+  activitiesDir,
 
   isBuildAllModule,
   buildModule,