+++ src/webpack.prod.js
@@ -1,4 +1,4 @@
-import webpack from 'es-webpack';
+import webpack from 'webpack';
 import ProgressBarPlugin from 'progress-bar-webpack-plugin';
 
 import config from './webpack.base';
@@ -11,4 +11,4 @@ const compiler = webpack(config, (err, stats) => {
   if (err) throw err;
 });
 
-compiler.apply(new ProgressBarPlugin());
+new ProgressBarPlugin().apply(compiler);